Add logic of run and restart container

+++ angular2cloud/dashboard/signals.py
@@ -2,33 +2,42 @@ import os
 import zipfile
 import shutil
 
-import docker
-
 from django.conf import settings
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.core.signals import request_finished
 
 from .models import Project
+from .utils import run_container
 
 #Business Logic
 @receiver(post_save, sender=Project, dispatch_uid="save_project")
 def on_save_project(sender, **kwargs):
-    path = os.path.join(settings.MEDIA_ROOT, kwargs['instance'].user.username, kwargs['instance'].domain)
-    dist_folder = os.path.join(path, 'dist')
-    dist_file = os.path.join(path, "dist.zip")
+    ''' This handler is called in tree ways:
+        1. When a new project is created.
+        2. When a new file is updated.
+        3. When the status is changed
+    '''
 
-    if os.path.isdir(dist_folder):
-        shutil.rmtree(os.path.join(path, 'dist'))
+    '''
+        In case of status is present in update_fields, this lines of code should not be called
+        because unzip or remove the dist folder is not necessary. update_fields is passed in views.ProjectActivate.post
+    '''
+    if kwargs['update_fields'] and 'status' not in kwargs['update_fields']:
+        path = os.path.join(settings.MEDIA_ROOT, kwargs['instance'].user.username, kwargs['instance'].domain)
+        dist_folder = os.path.join(path, 'dist')
+        dist_file = os.path.join(path, "dist.zip")
 
-    if os.path.exists(dist_file):
-        zip_ref = zipfile.ZipFile(dist_file, 'r')
-        zip_ref.extractall(path)
-        zip_ref.close()
+        if os.path.isdir(dist_folder):
+            shutil.rmtree(os.path.join(path, 'dist'))
 
-    if kwargs['created']:
-        client = docker.from_env()
-        name = "{username}_{domain}".format(username=kwargs['instance'].user.username, domain=kwargs['instance'].domain)
-        volumes = {dist_folder: {'bind': '/usr/share/nginx/html', "mode": "ro"}}
-        environment = {"VIRTUAL_HOST": "{subdomain}.{domain}".format(subdomain=kwargs['instance'].domain, domain="angular2cloud.com")}
-        client.containers.run("a2c-angular-nginx", name=name, volumes=volumes, environment=environment, detach=True)
+        if os.path.exists(dist_file):
+            zip_ref = zipfile.ZipFile(dist_file, 'r')
+            zip_ref.extractall(path)
+            zip_ref.close()
+
+        if kwargs['created']:
+            run_container(kwargs['instance'])
+        else:
+            if kwargs['instance'].status == 'RN':
+                restart_container(kwargs['instance'])
